+++ circuit/algorithms/src/lib.rs
@@ -21,8 +21,8 @@ pub use bhp::*;
 pub mod elligator2;
 pub use elligator2::Elligator2;
 
-pub mod keccak256;
-pub use keccak256::*;
+pub mod keccak;
+pub use keccak::*;
 
 pub mod pedersen;
 pub use pedersen::*;
